[Minor] alternative solution using minimal apis (WIP)
The app throws an error

InvalidOperationException: Body was inferred but the method does not allow inferred body parameters. Below is the list of parameters that we found: Parameter | Source --------------------------------------------------------------------------------- exchangeRateService | Services (Inferred) targetCurrencies | Body (Inferred) Did you mean to register the "Body (Inferred)" parameter(s) as a Service or apply the [FromServices] or [FromBody] attribute?